+++ packages/config/src/projects/other/zk-catalog/common/index.ts
@@ -1,0 +1,1 @@
+export * from './performedBy'
