+++ packages/config/src/projects/other/zk-catalog/common/performedBy.ts
@@ -1,0 +1,6 @@
+export const PERFORMED_BY = {
+  l2beat: {
+    name: 'L2BEAT',
+    link: 'https://l2beat.com',
+  },
+}
