+++ packages/config/src/projects/other/zk-catalog/index.ts
@@ -3,15 +3,9 @@ import { worldcoinsemaphore } from './worldcoinsemaphore'
 import { worldcoinsmtb } from './worldcoinsmtb'
 
 export * from './types'
+export * from './common'
 
 export const zkCatalogProjects: ZkCatalogProject[] = [
   worldcoinsemaphore,
   worldcoinsmtb,
 ]
-
-export const PERFORMED_BY = {
-  l2beat: {
-    name: 'L2BEAT',
-    link: 'https://l2beat.com',
-  },
-}
